% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aNigeria.R
\docType{data}
\name{aNigeria}
\alias{aNigeria}
\title{Event data slice from ACLED for Nigeria}
\usage{
data(aNigeria)
}
\description{
Event data slice from ACLED for Nigeria
}
\examples{
data(aNigeria)
head(aNigeria)
}
\references{
see acled and dorff et al 2020
}
\keyword{datasets}
