% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPreds.R
\name{getPreds}
\alias{getPreds}
\title{Calculate predictions using scenario accounting for estimation uncertainty}
\usage{
getPreds(modObj, scen, link = "logit", seed = 6886, sims = 100)
}
\arguments{
\item{modObj}{model object (glm or lm)}

\item{scen}{scenario object generated from scenBuild}

\item{link}{link function, options include: logit, probit (we'll build more in as we go)}
}
\value{
a data.frame object that includes columns for each of the treatment values
and then a final column with the predictions for that treatment. the number of
predictions generated per treatment correspond to the number of simulations (sims)
run.
}
\description{
This function generates predictions from model and inputted scenario
}
\keyword{predictions}
