% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{replScenVal}
\alias{replScenVal}
\title{replScenVal}
\usage{
replScenVal(scen, var, newval)
}
\arguments{
\item{scen}{a matrix object with column names}

\item{var}{a character argument that corresponds to a column of the matrix}

\item{newval}{a numeric argument giving the new value for the scen matrix}
}
\description{
This function helps users to change the values of
specific columns in a matrix
}
