% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleyhan_etal_2011.R
\docType{data}
\name{saleyhan_etal_2011}
\alias{saleyhan_etal_2011}
\title{Replication data for saleyhan_etal_2011}
\usage{
data(saleyhan_etal_2011)
}
\description{
Replication data for saleyhan_etal_2011
}
\examples{
data(saleyhan_etal_2011)
head(saleyhan_etal_2011)
}
\references{
see saleyhan et al 2011
}
\keyword{datasets}
