% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPreds.R
\name{getPreds}
\alias{getPreds}
\title{Calculate predictions using scenario accounting for estimation uncertainty}
\usage{
getPreds(
  modObj = NULL,
  beta = NULL,
  varcov = NULL,
  scen,
  link = "binary-logit",
  seed = 6886,
  sims = 100
)
}
\arguments{
\item{modObj}{model object (glm or lm), it's recommended to just supply values to beta and varcov}

\item{beta}{vector of regression estimates, usually obtained from coef of the model object}

\item{varcov}{variance covariance matrix}

\item{scen}{scenario object generated from scenBuild}

\item{link}{link function, options include: binary-logit, binary-probit, count, normal}
}
\value{
a data.frame object that includes columns for each of the treatment values
and then a final column with the predictions for that treatment. the number of
predictions generated per treatment correspond to the number of simulations (sims)
run.
}
\description{
This function generates predictions from model and inputted scenario
}
\keyword{predictions}
