% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenBuild.R
\name{scenBuild}
\alias{scenBuild}
\title{Helper to build scenarios}
\usage{
scenBuild(
  mData,
  ivs,
  scenType = "hypothetical",
  ivStats = rep("median", length(ivs)),
  treatVar,
  treatCategorical = TRUE,
  treatVals = NULL,
  intercept = TRUE
)
}
\arguments{
\item{mData}{data.frame object that contains at least the variables to construct the scenario}

\item{ivs}{character vector with all independent variables to be used}

\item{scenType}{type of scenario to construct: observed value approach or hypothetical}

\item{ivStats}{if hypothetical chosen then you have the option of specifying the summary stats to use}

\item{treatVar}{character vector indicating the treatment or treatments to be used}

\item{treatCategorical}{logical vector indicating whether each of the treatments specified is categorical or not}

\item{treatVals}{treatment values to test that should be wrapped in a list}

\item{intercept}{a logical indicating whether model was estimated with an intercept, default is TRUE}
}
\value{
a list, the first element being a list of scenario matrices, one for each
treatment value. second being a data.frame of the treatment values corresponding
to each scenario. third is a character indicating the type of scenario (hypothetical
or observed. last being a logical indicating whether or not the treatment variables
are categorical.
}
\description{
This function helps create scenarios for simulation analysis
}
\keyword{scenario}
